# rjweeks-securemessage-1.0.0-swagger.yaml
swagger: '2.0'
info:
  title: Secure Message API
  version: 1.0.0
  description: This is the RAS Secure Message API
  contact:
    name: Richard Weeks
    email: richard.weeks@ons.gov.uk
    url: https://github.com/ONSdigital/ras-secure-message
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
securityDefinitions:
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: http://uaa.ons.gov.uk/oauth/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: http://uaa.ons.gov.uk/oauth/auth
    flow: accessCode
    scopes:
      write: allows modifying resources
      read: allows reading resources
security:
  - accessCode:
    - read
...
schemes:
- https
paths:
...
  /message/{id}:
    get:
      parameters:
      - name: id
        in: path
        description: Message ID
        required: true
        type: integer
        format: int64
      - in: query
        name: reportingUnit
        description: Reporting Unit
        required: false
        type: string
      - in: query
        name: survey
        description: Survey name
        required: false
        type: string
      tags:
      - Respondents
      - Respondent Liason
      summary: Fetch message by id
      operationId: getMessage
      description: Fetch message by id
      produces:
      - application/vnd.collection+json
      responses:
        200:
          description: Returns message by id
          schema:
            type: array
            items:
              $ref: '#/definitions/Message'
          examples:
            application/vnd.collection+json:
              msg_to: richard.weeks@ons.gov.uk
              msg_from: respondent@xyz.com
              body: 'Return message details by ID'
        400:
          description: Bad syntax
        403:
          description: Forbidden
        404:
          description: Message not found
    put:
      parameters:
      - name: id
        in: path
        description: Message ID
        required: true
        type: integer
        format: int64
      - name: read
        in: query
        description: Mark message read or unread
        required: false
        type: boolean
      tags:
      - Respondents
      - Respondent Liason
      summary: Update message status by id
      operationId: updateMessage
      description: Update message status by id
      produces:
      - application/vnd.collection+json
      responses:
        200:
          description: Updated message by id
        403:
          description: Forbidden
        404:
          description: Message not found
...
definitions:
  Message:
    type: object
    required:
    - msg_to
    - msg_from
    - body
    properties:
      msgId:
        type: string
        description: Message identifier (UUID)
        example: f9419ea4-029e-40cd-bd9c-115533dae8c5
      threadId:
        type: string
        description: Thread identifier (UUID)
        example: 212faf46-931f-4170-9b96-949e20722126
      msg_to:
        type: object
        required:
        - id
        properties:
          id:
            type: string
            format: urn
            example: urn:ons.gov.uk:id:respondent:001.234.56789
          emailAddress:
            type: string
            format: email
            example: "richard.weeks@ons.gov.uk"
          firstName:
            type: string
            example: "Richard"
          lastName:
            type: string
            example: "Weeks"
      msg_from:
        type: string
        description: Message sender
        example: Andrew Torrance
      subject:
        type: string
        example: Survey enquiry
      body:
        type: string
        format: byte
...
host: virtserver.swaggerhub.com
basePath: /rjweeks/securemessage/1.0.0